Fixing SWAN serial/parallel compilation issue. Adding gitignore.
SWAN runs a script to pre-process files as part of the build process.
This process is different for serial and parallel modes. The process
needs to be cleaned if changing from a serial compile (adcswan) to a
parallel comple (padcswan). The makefile now keeps track of this via a
hidden file .swanlastcompile.

The circle.yml file has been updated for SWAN and XDMF compiles. An
issue was fixed with the XDMF library locations. It now expects the
standard directory tree (lib/include/share/bin). Use XDMFHOME to specify
the location of the directory just as you would with NETCDFHOME. The
HDF5, netCDF, netCDF-Fortran, and XDMF libraries are now built by the
build server since it was unable to use the pre-installed versions.

ASWIP did not compile with XDMF enabled. This has been resolved.

Finally, a .gitignore has been added to keep track of files that are
not supposed to be included in the repository but might be included by
mistake.