Handling version_cmake.F file when conducting Windows Native build
The windows native build will assume that it does not have to
generate the version file itself. Instead, it will fall back
to a version file made available in the main source directory.

By default, it looks for version.F, however if not found it will
fall back to version_default.F.