Remove requirement from fort.14 file that node and element numbers must be consecutive integers from 1 to NP and 1 to NE, respectively.
Within the fort.14 file, node and element numbers can now be any
arbitrary number, as long as they are unique integers and as long as
the mapping of elements to nodes remains valid. This aids in the
development and testing of subdomain modeling as well as in mesh
editing. Internally, only the reading of the fort.14 file has been
modified, maintaining a list of 'node labels' that will be used
when writing output. Output files are written using the node numbers
as defined in the fort.14 file.