Fixing sparse output format for serial runs
The sparse output format was not being correctly written for
serial runs. The output file was blank except for the header
for each time snap.

Also, removing the "-DDEBUG" flag in favor of "-DEBUG". (i.e. ifdef EBUG)
This avoids a situation where the compiler defines variables
within the code to "1", which is not the intention