Fixing EtaDisc initialization for type 52 flux boundaries
The flux settling time is used to allow the river boundary conditions to equalize before beginning other forcing in the model. It does not appear that it is a hard requirement that it be used, but it is a good practice.

In recent versions of the code, the `EtaDisc` variable has been changes so it is initialized to `-9999.0`. However, if the flux settling time is not used, this value begins to find its way into the equations used to generate the `GWCE_LV` which causes fatal instabilities in the first few model time steps.

Previous versions of the model initialized this value to 0.0, though a more appropriate way to do this seems to be to initialize it to `ETA2` once the geoid offsets and `river_above_msl` attributes have been applied so it is consistent.

This commit initializes `EtaDisc` to `Eta2` during cold start. This does not appear to be necessary at hot start since those values will be read from the hot start file. This commit fixes #79.