Fixing logic when checking for an eta_max type boundary change
The code was originally set up to first check if
either side of the weir was active and then check if
either side the weir had an elevation that could trigger
the boundary to change elevation. The check against nnodecode
needs to occur at the same time as the check against eta2.

The problem becomes when one side of the weir is active and the
other side has a topographic elevation greater than the weir
change criteria, the code will sense this incorrectly because
the active side of the weir will trigger the first check and the
topographic elevation will trigger the second, even though the
node is not active. This change performs both checks simultanously
so that this is no longer an issue.