Fixing bug introduced when hashtable module was introduced
Hashtable module previously did not account for parallel
simulations. ADCPREP and PADCIRC have been adjusted to
use hashtable. This fixes issues in max outputs as well as
timeseries output files.

However, for parallel simulations, node numbering is still
enforced to be sequential via adcprep. Once the data
is in the subdomains, the restriction is lifted.

Finally, note that subdomains now have the labels from the
global nodes and not a new set of indicies. This is carried
for the fort.13 and fort.14, however, other files currently
still use indexing.