Adding check for intel compiler and applying heap-arrays option. Fixes #100.
The `-heap-arrays [stacksize]` option for the intel fortran compiler
forces Intel to use the heap for large implicitly created arrays. This
is the default setting for most other compilers. The `work/makefile` has
been modified to print a message warning the user to add this option and
the CMake build system has been modified to add this setting automatically
by querying the current machine's stack size via `ulimit -s`. This option
should also be added to the netcdf-fortran Fortran compiler flags as this
is usually where this segmentation fault can occur. The option was added
to the ADCIRC build process to prevent a simliar issue.

Additional changes were made to process the `NETCDFHOME` and `XDMFHOME`
environment variables directly if they exist after processing the predefined
variables that would exist from a cmake command line option, specifically
`-DNETCDFHOME` and `-DXDMFHOME`. The command line options are given priority.
If both of these cases fail, the user is prompted for the location of the
libraries before the options are allowed to be enabled.