Write alternate value (-99999) to two-record output files in serial runs
In a serial ADCIRC run, the alternate value of -99999 is not used to
indicate that a node is dry in output files that contain two records
(e.g. fort.64). Instead, the "initial_value" is used, which in the case
of velocity output is 0.0, implying that the node is wet but water
velocity is zero.

This commit fixes this issue by using the supplied alternate value
instead of the initial value.