Extended lon to 4 columns for NWS19GET and NWS20get
In an email with subject "How to modify fort.15 when NWS is changed from 8 to 19，very thanks", a user reported issues in getting NWS=19 or NWS=20 to function properly when simulating in the western Pacific.  I identified part of the issue as being related to ADCIRC assuming 3 columns for the longitude, rather than 4, which is permitted by the format since it should otherwise be a space character.