Halt run if offsetTimeIncrement.LE.0
I found that setting the input offsetTimeIncrement to zero (as one might do in the case of a constant offset run) and ending the offset input file with a "##" line (this is the file delimiter) will lead to division by zero and so NaNs.  This solution was substantially less complicated than permitting the zero value and trying to figure out at run-time whether the file had actually ended or whether the "##" was preceding the next snap of data.