Using a negative to indicate owi filenames should be read
A negative numSets in the fort.22 allows the user to specify
an unlimited number of oceanweather domains from coarse to fine.
A practical limit is 99 oceanweather domains since presently
the code needs to calculate the iounit on its own. In the future
the iounits throughout the code should be calculated dynamically