Adding option for an elemental subgrid barrier
The nodal attribute "subgrid_barrier" specifies the
height of a subgrid barrier and optionally friction to use when
overtopping. This attribute will shut down flow across
an element until the subgrid elevation is exceeded. When the elevation
is exceeded, flow will be allowed to pass between elements normally
using the subgrid barrier specified elevation

Because this is an elemental attribute, the three nodes that make
up an element must be specified in order to trigger this.The water
surface elevation used is the max of ETA2 for the three nodes.
Additionally, all three nodes must specify a valid
subgrid_barrier in order for the option to be used.