Fixing compile under gnu compiler suite >= version 10
The type mismatches in the arguments are fixed for the adcirc
side of the house, however, the SWAN edits are much more extensive
and should be edited on their end rather than ours. To allow the
code to build, cmake will add the flag "-fallow-argument-mismatch"
when compiling with gfortran > version 10 to only the SWAN source
files.

Fixes #238