Add global dependencies required by allMessage
Calls to allMessage on lines 424 and 703 of subroutine coldstartMeteorologicalReopenInputFile require the use of allMessage and DEBUG from the global module. These dependencies are now brought in at the start of the subroutine (with the addition of line 393).

Likewise, similar calls to allMessage on lines 781 and 1229 of subroutine hotstartMeteorologicalReopenInputfile also invoke the aforementioned dependencies, so these are also now brought in at the top of this subroutine (the new line 742+743).