Fix for race condition in "reopenLogFile"
When running in parallel, subroutine "reopenLogFile" in global.F induces a race condition with the use of "inquire" to check whether a file exists. Replace the "inquire" with "open" using the "STATUS='OLD'" specifier. If the file does not exist, then the error flag returned by the "IOSTAT=" specifier will indicate as such, and we can handle this as needed. If the file is already opened, then this approach will not cause an error. https://gcc.gnu.org/bugzilla/show_bug.cgi?id=92836