Potential correction for incorrect STATUS= specifier in openLogFile
When running in parallel, open statements in routine openLogFile in global.F cause the error "incorrect STATUS= specifier value for connected file". Possibly caused when the file attempting to open already exists. To work around this, we first use an open statement with the "STATUS='OLD'" specifier. If this returns an error flag, then we use another open that has the "STATUS='REPLACE'" specifier.