Addition of capability to simulate effect of an internal barrier with a pump
The code has been modified to add an additional flux-specified internal boundary that simulates an internal weir with an embedded pump. The functionality is similar to the weir/pipes internal boundary, but with a user-specified flow/width (PUMPQ) that is in one direction only (front face to back face). The pump is turned on when the water level on the front face exceeds a threshold (PUMPONWL) and is turned off when the water level on the front face becomes lower than a threshold (PUMPOFFWL). This internal boundary condition can be implemented either as an essential  condition with free tangential slip (IBTYPE=6) or as a natural condition with free tangential slip (IBTYPE=26).

Rebasing from development branch.