Slope+Wind Limiting
Enforce slope limiting on the water surface gradient when the component of water surface gradient which lies in the direction of the bathy gradient exceeds a user-specified limit. This limit is specified in a namelist &limitControl as the first attribute, slim.
Second in this namelist is the logical parameter windlim, which allows the user to toggle wind force limiting. If windlim = true, then in shallow areas, e.g., any node where total water column height is between 3*h0 and h0 (h0 is required in the fort.15 file), any wind forcing is ramped from 1 to 0 by a tanh scaling function.