Initialize Slope Limiter to Off
Initialize limit for slope to a very large value to effectively switch it off, unless the user specifies a limit in the limitControl namelist in control file (fort.15). This was previously initialized to 0, which provided the opposite effect of this intent, i.e., infinite slope limiting was enforced and no elevation gradient was computed. Now, there should be no slope limiting by default.