Introduce TK averaging to wetdry routine
Whenever a node is turned from dry to wet, the value of the friction coefficient TK at the source node is reset. If a wet node is the source node for more than one dry node, then the value of TK was overwritten in previous versions of the code. With this update, the number of times a node is the source for turning another node from dry to wet in a single timestep is counted and the resulting values of TK are averaged at the source nodes.