Remove erroneous update on TKM
The scheme for updating the TKM array during parallel runs was incorrectly implemented and caused erroneous internal tide dissipation in the open ocean, due to looping over all nodes and not just the ones involved in wet/dry transitions. Remove this update for now, until a more appropriate scheme is implemented.