Support for directly computing Vel in wetting/drying
Adding the namelist "velwdControl" and the corresponding (logical) parameters "directvelWD" and "useHF". The option "directvelWD" lets the user choose to compute the flow velocity in the wetting/drying algorithm directly from the gravity - friction force balance. The previous code uses the user chosen value of "velmin" to approximate the velocity, and then checks this velocity against the very same value of velmin to determine whether to wet a node. The directvelWD option allows the user to avoid this circular logic and obtain a more physically realistic approximation to the velocity, which we then check against velmin to find out if a dry node should be wet. We also provide the option "useHF" to invoke a larger approximation of friction in the wetting/drying algorithm, wherein if useHF = true, then we assume the friction coefficient Cf = 1 in the computation of the velocity due to the friction balance. On average, this increases friction along the wetting front considerably and can significantly slow the propagation of the wetting front during a flooding event, should the user deem this necessary. More information on the above can be found in the notes here: https://docs.google.com/document/d/1-OW2ivoSWmGX--QUd7VrbE7SCiTCQFxACWPhICm1HWk/edit?usp=sharing

We also include in these changes a simple fix to the writing out of the "rotation matrix" in the "mesh" module. Before, an array temporary was created for the write statement, which is placed on stack memory. Now, we use a dummy array to hold the values being written out and use this array in the write statement, so that the array is placed in heap memory instead.