feature: use migratable LUNs to support AMPI load balancing
This scheme adds a lun registration call after OPEN when a
numerical LUN is in use.  This will then trigger the automatic
closure of LUNs before migrating a rank and reopening those LUNs,
at the prior file offset, after migration.  When combined with
PIEglobals, this provides complete support for migration
for code making direct use of Fortran I/O.