MPI Updates for Lateral Stress Tensor
Adding MPI updating for components of the lateral stress tensor. This resolves slight differences that can arise at subdomain boundaries when using a 2-part lateral stress formulation during a parallel run. Experiments show that not updating these can cause the precision of a solution to approach single precision, when the code is designed to maintain double precision.