Adding compiler and cmake warnings for model behaviors enabled at compile time
This change ensures that the input files the model uses generate
results with the same assumptions and different builds do not change
physics or input file formats.

Flags being deprecated:
-DPOWELL - Enable the Powell wind drag formulation. Users should
switch to using the &metControl namelist.

-DWARN_ELEV_DEBUG - Enable writing of the fort.69 file. Users should
switch to using the &warnElevControl namelist.

The code uses the #warning compiler directive to alert the user during
the build and the cmake build system will also throw a warning alerting
users to deprecation at configuration time