Changes for repartitioning option in adcprep
Allows the user the option to repartition a grid that has already been partitioned in two, e.g., into wet and dry zones. 

User selects option "8" to use the new feature. Reads in a partmesh.txt file with 1 and 0 mappings for each node, that the user creates themselves before running the code.