Introduce Global Variables H1 and H2 + VELCHECK compiler flag
Use global variables H1 and H2 to save the flow depths at the current and previous timestep, so that if bathy is changed (either due to subgrid barrier or time varying bathy) the correct value for H1 is retained. Also prevents us from having to compute DP + ETA as much.

Also, place checks for velocity errors (either NaN or very large values in the soln) behind the "VELCHECK" compiler directive.