Adding wave stress gradient limiter
Final changes for adding option to use a wave stress
limiter within ADCIRC. Note that when wave stress
gradients are limited, the values outputted in
the rads.64 are not limited, thus representing the
original values provided to ADCIRC. The wave stress
gradient magnitude limiting is by default turned off.
There are two controls for this new feature, both are
contained within the waveCoupling Namelist. The first
is Limit_WaveStressGrad which is a logical variable that
is set to .false. by default and means that limiting is not
used, setting it to .true. turns on the checking for limiting
of wave stress gradient magnitudes. The second variable is
WaveStressGrad_Cap which is a real number value with a default
setting of 1000.0d0. This value represents the upper limit to the
size of wave radiation stress gradient magnitude. If the
magnitude is greater, then the rsx and rsy values are scaled
proportionately so that the magnitude is no more than the
WaveStressGrad_Cap. The default value is set to an unrealistically
high value, in the case limiting is inadvertently turned on. More
reasonable values for the upper limit would be in the range of 1.0
to 0.001, but should be tested for sensitivity to results.  Finally,
bug fixes were made to allow rads.64 file creation when using NRS=1
and code changes made to allow for rads.64 output whenever NRS>0. TCM v56 2024