Adding test builds using Rocky Linux and gcc compilers
The version is the system default for the Rocky Linux distribution
which is GCC 11.4.1. Rather than install another version via
spack, this gives us a little flexibility to use different operating
systems. The current intel based compilers are installed on a ubuntu
based container while these gcc compilers are from a redhat/rocky
based container

This adds another compiler to our list of compilers that
are automatically tested. Currently tested are:
1. Intel (LLVM, i.e. ifx, icx, etc)
2. Intel (Classic, i.e. ifort, icc, etc)
3. GCC (i.e. gcc, gfortran).