Update writer.F
Added a check to see if the labels_g was already allocated or not. When running in coupled mode via cstorm, every time adcirc pauses, the writers pause but labels_g is not destroyed. Thus when the second time the coupled model entered into adcirc_run and writers were being used, the allocate (labels_g) caused the model to crash.