Using pre-computed array for csr matrix position
Previously, the code computed these positions each pass
through the gwce when the consistent matrix was used. Now,
the code computes these positions as part of the initialization
and stores them in an array. The trade off is that you're allocating
an array that is ~6x the number of local elements. This is balanced
by not needing to perform a loop with 6 branches 8 times per time step.
In both cases, there is a multidimensional array lookup, so the cache performance
seems similar.