Added option to remove dphidt from computations
The dphidt term came out of the averaging of the GWCE. At high timesteps, this term can because large negatively and violate the hyperbolicity criteria of the GWCE causing a failure.

Added an item to the subgrid namelist 'phi_time_derivative' that when set to true will turn off this term allowing for higher timesteps. By default it is off.