Namespacing ADCIRC cmake options and adding strict compile option
This allows ADCIRC to be included as a subproject in other
cmake projects

The strict compile option turns all warnings on when building
the code. Currently, only the .F90 files are enabled since those
newer files are being built with this in mind.

Updating files so that they compile with the strict options