Removing behavior where stormnumber controlled the wind multiplier
This behavior seems to be from an old forecasting system where the
wind multiplier was controlled by storm number. This is extremely
unintuitive and undocumented, and therefore it has been removed.
A WindMultiplier namelist parameter has been added in its place, however
it is currently only activated for the holland model and it does
not modify CLE15 wind speeds