Adding developer mode which builds code with additional flags
Particularly for GCC, the additional flags can help diagnose
find certain issues in the code. The GCC flags that are added
in developer mode are:

-Werror
-Wall
-Wextra
-Wconversion
-pedantic
-fimplicit-none
-Wuninitialized
-Wsurprising
-Wuse-without-only
-Wimplicit-procedure
-Winteger-division
-Wconversion-extra

Note that these flags are only added to *.F90 files. This is to
make the changes required by these flags incremental because they
would trigger many errors in the older portions of the code.