Using standardized datetime module and datetime-aware logging in the code
This work centralizes the calculation of dates and times into the
fortran datetime module (external dependency) that is included
with the adcirc source package.

The updated logging messages use the netCDF base date to compute the
time-aware status logs in the code.

The status logs will now also provide an elapsed/remaining time
estimate to the user.

A custom MPI type has been used to replace the previous
WarnElevSum function. Previously, this was an allreduce on an integer
and now the allreduce is performed on a custom type which allows
the status logs to write the maximum elevation and velocity values,
the node, and the processor where these values occur. This gives
the user better insight into the model status than previously
where only processor 0 could log its information to the screen.
This also generalizes the checks on velocity that were done previously
using a duplicated code block.

Lastly, ensuring that a two digit year is interpreted as yr 2000
rather than year 0