Fixing variable aliasing. Conversion of owiwind_netcdf.F -> nws13.F90
The WTIME2 variable was called as both an intent(in) and
intent(inout) and was aliased to two variables within
the NWS13GET function. This could lead to misoptimization
by the compiler where memory was changed in unexpected
ways

Additionally, moving to the F90 style for nws13 module.
This involves the module renaming and adherence to
string compiler flags/formatting