Removing custom text->numeric parsing code from global.F
The code now relies on standard fortran parsing using the read
function. The old code could (very, very) slightly shift stations
because of numerical precision issues in the parsing routines.
The new routines are robust to this issue.

Fixing bug in 'parse' function in global.F which was obscured
by the parsing functions. The bug occured when the delimiter was
left at the beginning of the string instead of being properly
handled