Using RAII based logger to set/unset the message sources
ADCIRC followed the pattern to generate a call stack using
a call to setMessageSource at the start of a function and
a call to unsetMessageSource when it left the function. This
is fine, however, a mismatched pair of calls could cause the
code to walk off the front or back of the messageSources array.

This means that the user needs to correctly annotate every entry
and return path for a function or risk a bug. What this work does
is employ the compiler to handle these tasks. The user specifies
a call to a macro:

LOG_SCOPE_TRACED("myfunction", trace_condition)

This automatically instantiates a variable with type t_log_scope
and makes the calls to setMessageSource. On its own this isn't a
huge improvement (apart from the fact that the boilerplate of
calling "call allMessage("Enter")" is removed), however the nice
part about this macro is it will call the destructor when the
variable goes out of scope. This means that there is no need to
annotate return paths and the compiler will handle this correctly

This has been tested with gcc 8 and above, Intel compilers, and
nvhpc compilers. Compilers adhering to the Fortran 2008 standard
(aside from coarray Fortran) will handle this correctly.