Removing mpi use lists for legacy module
The `use mpi, only: xyz` syntax is not supported by
some mpi libraries and will cause an error when
importing the MPI functions using the legacy module.

Long term, we need to move to `use mpi_f08` since
this is the modern and recommended module to use.
`use mpi` is only provided for legacy compatibility.

This work removes the `use mpi, only: xyz` from
the code as a workaround for these mpi libraries