Fixing issue where certain compiler/mpi combinations hang within SWAN
SwanPunCollect.ftn90
SWAN issues an MPI_GATHER command which should collect data to process 0,
however, if the mpi implementation makes assumptions about the receive
buffer on non-root processes of the MPI_GATHER, then the code could
be working with uninitialized memory. Instead, we allocate the icount
array on all processes and set to 0 to explicitly initialize it on
all processors which resolves the issue.

The check is also updated to use MPI_ALLGATHER. This is important
because if the check for node count did fail, only processor 0
would ever know about it introducing a second way for the code to
stall